Remove Ahem font from canvas-TextMetrics-width test
A test in the fast/canvas directory tried to used the Ahem font to test
the width value obtained from using `ctx.measureText()`. This font, as
explained in https://web-platform-tests.org/writing-tests/ahem.html,
consists of only squares and rectangles of specific dimensions.

The test in question expected in its reference image the for the text
"Hello World" to be rendered, which indeed happened as the commands were
executed without waiting for the font to load. Therefore, the fallback
font, "serif" in this case, was used. This change removes the Ahem font
from the test and just replaces it with "serif".

Bug: 341213359
Change-Id: I4ad178c781e004c7f3bbbd066d8823c403a886b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5783320
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Commit-Queue: Andres Ricardo Perez <andresrperez@chromium.org>
Reviewed-by: Jean-Philippe Gravel <jpgravel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1347535}
NOKEYCHECK=True
GitOrigin-RevId: 3a4fef86522ecf84fe56d25b6fd4116083424dbb