Fall through to <general-enclosed> in @supports parser
Correct <general-enclosed> behavior in the (streaming) @supports
parser has been practically impossible due to lack of restarts.
For example, for "@supports selector(/* ... */)", if we encounter
something invalid within the selector(), the BlockGuard skips
to the end of the block, and all we know is that "this didn't parse
as a selector". Previously, we could not go back and check if
it's valid as <general-enclosed>.

As of CL:4239860, the streaming parser gained the ability to restart,
and this CL makes use of that to fix the <general-enclosed> behavior.
The restarts are done via RestoringBlockGuard, which acts like
BlockGuard, but restores to a previous state (at the start/outside
of the block) instead of skipping forwards.

Fixed: 40237918, 40737031
Change-Id: Id8e659620b7a8da859a01c46567f7c6b765310a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5811736
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1347543}
NOKEYCHECK=True
GitOrigin-RevId: aec6ca965e8dcda3f10654c5576afd5559855d3c