Use direct shaping in TextMetrics
Direct shaping is already being used lazily for the new methods added as
part of extended text metrics. It wasn't done directly when
`ctx.measureText()` was called to make sure its performance was not
affected, as it already shapes in a different way when calling
`Font::Width()`. This means we are effectively shaping twice: once when
`ctx.measureText()` is called, and once when `ShapeResult` objects are
first required.

Since shaping is the most expensive operation for `TextMetrics`, it is
useful to have the objects right away without having to shape once again
when the new methods are called.

This CL shapes directly from the moment `measureText()` is called,
provided that the `Canvas2dTextMetricsShaping` feature is enabled. The
flag for this feature is enabled by default, and is meant to act like a
kill switch in case we need to undo the change for performance
regressions.

The performance between the two approaches was tested locally, and for
the relevant test cases performance seems to be unaffected. The
scenarios that play into the old method (that uses `CachingWordShaper`
under the hood, which is planned to be deprecated) are strings with
several space separated and repeated words.

Since the former approach ignored cross-space kerning, a pixel test was
checking with a wrong width. It was rebased.

Bug: 341213359
Change-Id: I881156c9fd03f0f1f490bd1088f706c9d7224f3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5758312
Reviewed-by: Jean-Philippe Gravel <jpgravel@chromium.org>
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Andres Ricardo Perez <andresrperez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1347553}
NOKEYCHECK=True
GitOrigin-RevId: c3eb0244cca8371bae6c830149a0db94c79a48e6