Fix web test
This test fails because it would often fail to receive the
expected scroll events. This seems to have been happening because the
test would issue a scroll gesture before the compositor was ready.
This readily reproduced locally and appears to be fixed by adding
waitForCompositorCommit().

Let's watch the test results[1] for an improvement in the flakiness
before removing the TextExpectations entry.

[1]https://ci.chromium.org/ui/test/chromium/ninja%3A%2F%2F%3Ablink_web_tests%2Fvirtual%2Fthreaded-prefer-compositing%2Ffast%2Fscroll-snap%2Fanimate-fling-to-snap-points-2.html?q=V%3Abuilder%3DLinux%2520Tests%2520%28dbg%29%281%29+V%3Aos%3DUbuntu-22.04+V%3Atest_suite%3Dblink_web_tests

Bug: 1186753
Change-Id: I9ef8e1a33ae5a7970102ed9a0a6bb285bc53959f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5817693
Commit-Queue: David Awogbemila <awogbemila@chromium.org>
Reviewed-by: Kevin Ellis <kevers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1348028}
NOKEYCHECK=True
GitOrigin-RevId: 84ea0165d1ed5e4d26e2e15c0c396f292b05e14d