[Interop] Fix coalesced/predicted event targets in trusted events.
Our past attempt [1] to fix the targets in untrusted events
accidentally reset the targets in trusted events!  This CL fixes the
bug and updates the WPTs.

The change is still in "experimental" state.  We will flip it to
"stable" in a separate CL.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/5709870

Bug: 353538500
Change-Id: I167b72d5df588fae11eb66206a5190226020247e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5812458
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Mustaq Ahmed <mustaq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1348055}
NOKEYCHECK=True
GitOrigin-RevId: 365b4154cbdd41b86efd93c898a8350c13a27c13