[gpu] Allow sync readback through CopyToGpuMemoryBuffer path
Allow synchronous readback through CopyToGpuMemoryBuffer path on
all platforms which was earlier enabled only for Windows. This will
allow clients like Meet WebRTC to use readback for the BreakoutBox
feature on Linux and can help with the migration. This enables the
path to be used on platforms where native GpuMemoryBuffers are not
available.

Bug: 347005844
Change-Id: Icbc5922fedfd173fec884ae9a012281562b6fec3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5812444
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Saifuddin Hitawala <hitawala@chromium.org>
Reviewed-by: Brendon Tiszka <tiszka@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1348173}
NOKEYCHECK=True
GitOrigin-RevId: da1e89c574800c75add190411a48bc16bc525361