Deflake test for touchmove event being passive
Replaces waitFor with a waiting on a promise for a touchmove event.
The waitFor method is know to be the source of testing flakes since
using a frame based timeout, and frames are pumped as quickly as
possible in our testing framework.

Deferring update of the TestExpections to a follow up CL once the
fix has been verified.  Unable to repro the flake locally with the
fix.

Bug: 936891
Change-Id: I531502526669f5074d49b66e19e77c657c4a80f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5824011
Reviewed-by: Sahir Vellani <sahir.vellani@microsoft.com>
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1348189}
NOKEYCHECK=True
GitOrigin-RevId: ee7082e77fad67c07f9da8ad8ff54f101fc5a1e6