[discard] Address same-document navigation crash during discard
This CL clones content::DocumentState / WebDocumentLoader::ExtraData
and passes it to document loaders of empty documents replacing
the current document as part of a discard operation or JavaScript URL
navigation.

DocumentState is cloned instead of extracted as the original document
state object may be needed to process operations in the replaced
document's unload handlers (such as same-site navigations, see linked
bug).

DocumentState::navigation_state_ is not cloned as this is specific to
the document being replaced and should not be moved to the new empty
document instance.

Bug: 361658816
Change-Id: I9d39867a1688e81b17e18e5f103f53d63aed96fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5821193
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1348293}
NOKEYCHECK=True
GitOrigin-RevId: af57eb3e1a26f8015fe3e2783255bf10d4fb2d84