Option for JavaScriptExecuteRequestForTests() to ignore content settings
Prior to this CL, the javascript content setting was always honored by
EvalJs. This CL makes EvalJs ignore the content setting by default, with
a flag to honor the setting.

This CL is a copy of ps#4 from
https://chromium-review.googlesource.com/c/chromium/src/+/5350575

In blink::LocalFrame::JavaScriptExecuteRequestForTests(), add another
option for whether JS execution will honor content settings or not. Then
plumb it down to LocalFrameMojoHandler that does the actual work to
select the appropriate ExecuteScriptPolicy, and plumb it up to
content::EvalJs() so callers can ignore content settings if they need to
do so.

This is useful for content settings tests, where JS content settings
need to be disabled for a frame, yet the test itself wants to execute JS
in the same frame.

Change-Id: I239840db38c6cba0d9bf2b5ff8c1764743a33ee7
Bug: 40282541
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5806109
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Richard (Torne) Coles <torne@chromium.org>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Commit-Queue: Erik Chen <erikchen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1348323}
NOKEYCHECK=True
GitOrigin-RevId: 40f6382d9269b2ee7c877a1bfa2dc3302584c893