Stop styling links as :visited in Fenced Frames
When one of the two partitioning flags is enabled, we no longer want to style any links in a Fenced Frame as :visited. This prevents :visited state from other frames from leaking into this isolated context.

For more information on how we arrived at this decision, please see this internal memo: https://docs.google.com/document/d/1G0DnZd7dvm7Ot1wjzYBorwlTgdURra-7f30woX02QRk/edit?usp=sharing

Summary of the above document for external use: Partitioned :visited links depend on two things: (1) what we store and (2) what we style. For Fenced Frames, we do not store any navigations that take place in a Fenced Frame tree in the VisitedLinkDatabase. Thus, we cannot display any of those navigations as :visited in any frame. HOWEVER, we do continue to store and style navigations that happen in other non-fenced frames. As a result, if the triple-key of <link url, top-level-site, frame-origin> for a given Fenced Fame matches a previously stored non-fenced frame navigation, it will still show that element as styled. While this is not a regression, it does not match with the isolation model associated with Fenced Frames. So together, the Fenced Frames team and the Visited Links team have decided to disable :visited-ness in all Fenced Frames for the time being.

Bug: b/353585049
Change-Id: I92e3aa07578726248dbba5c75e979083c874ec0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5781774
Reviewed-by: David Baron <dbaron@chromium.org>
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Commit-Queue: Kyra Seevers <kyraseevers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1348583}
NOKEYCHECK=True
GitOrigin-RevId: 746e5806dd709fd9bfb6dfbde87218b02d2b72f8