Make @import prelude parsed using the streaming parser.
There is one exception still: If there's a media query in there,
it is still parsed using the non-streaming parser. However, the
supports part and the prelude in general are parsed streaming.

Note that this fixes a bug (also in the tests) where empty supports(),
or supports() with an invalid declaration inside, would be allowed;
there doesn't seem to be any allowance for this in the grammar
(<general-enclosed> is not included in the @import grammar, only in
@supports).

Bug: 1500904
Change-Id: Ida837291feac344e847b0e3982acff9dd73c7d53
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5822594
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1348630}
NOKEYCHECK=True
GitOrigin-RevId: 27b26836d1d41c6dfa4075ecb75aa34da2b920e0