Fix 304 responses handling in ResourceLoader
After crrev.com/c/362788339, ResourceLoader::DidReceiveResponse()
doesn’t handle the body of 304 response for revalidation request. This
is causing the 304 resource loading failure. This happens when the
disconnection of the passed body pipe is handled in the network service
as if the request was canceled, before the network service sends the
URLLoaderClient's OnComplete IPC.

To fix this issue, this CL changes the DidReceiveResponse() method to
keep the body handle in the ResourceLoader.

Bug: 362788339
Change-Id: Ic0e5e5d41965940ebc6239291dce0c768dfdbb9a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5822737
Owners-Override: Srinivas Sista <srinivassista@chromium.org>
Reviewed-by: Brandon Maslen <brandm@microsoft.com>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Commit-Queue: Kouhei Ueno <kouhei@chromium.org>
Commit-Queue: Prudhvikumar Bommana <pbommana@google.com>
Cr-Commit-Position: refs/heads/main@{#1348877}
NOKEYCHECK=True
GitOrigin-RevId: ffd48736e9e90e833af3bb5ba6d42faa797f77be