JPEGImageDecoder::GetGainmapInfoAndData: Don't extract gainmap data
Don't extract the gainmap as a separate SegmentReader in
JPEGImageDecoder::GetGainmapInfoAndData. Instead, move this operation
to JPEGImageDecoder::OnSetData.

This allows us to rely on ImageDecoder::aux_image_ to determine if
we are decoding the base image or the gainmap image. For JPEG, because
libjpegturbo isn't gainmap aware, and so we will need to do this
explicit extraction. For AVIF, we will not.

This is a step towards removing ImageDecoder::GetGainmapInfoAndData,
and having cc::PaintImage have a single cc::PaintImageGenerator for
both the base and gainmap image (instead of one for each), and adding
a cc::AuxImage argument to cc::PaintImageGenerator::GetPixels (et al).

Bug: 356827770
Change-Id: Ie80076ae67aff524856090779efedd87f2caba4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5812388
Commit-Queue: Urvang Joshi <urvang@chromium.org>
Reviewed-by: Urvang Joshi <urvang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1348964}
NOKEYCHECK=True
GitOrigin-RevId: 51c72fe723c583c4beabc0c66e98b8fccd0e85df