[Canvas] Remove unneeded call in Canvas2DLayerBridge::NewImageSnapshot()
This method does the following:
* Calls GetOrCreateCanvasResourceProvider(), returning null if it is
  null
* Calls host_->FlushRecording()
* Calls GetOrCreateCanvasResourceProvider(), returning null if it is
  null
* Calls ResourceProvider()->Snapshot()

As detailed in [1], the only impact that FlushRecording() can have on
the resource provider is to potentially turn it invalid. The second call
to GetOrCreateCanvasResourceProvider() will be a functional no-op
regardless of whether the resource provider is valid or invalid, and
ResourceProvider() will return the resource provider regardless of
whether it is valid or invalid.

3 of the 4 subclasses of CanvasResourceProvider return null for
Snapshot() if the instance isn't valid, and the 4th
(CanvasResourceProviderPassthrough) always returns true for IsValid().

Hence, the second GetOrCreateCanvasResourceProvider() call here serves
no purpose and can be removed.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/5824628

Bug: 40280152
Change-Id: Ief7388f4507347e724d21412c52e2928014630c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5826513
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Reviewed-by: Jean-Philippe Gravel <jpgravel@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1349073}
NOKEYCHECK=True
GitOrigin-RevId: d0c6926b4d7b073656b7461799270ba62daafec6