[Editing] Paint highlights for first-letter text fragment
When highlight markers are set on a layout text with first letter
pseudo element, the associated node for first-letter part is not
invalidated for paint and highlights for markers like text match,
spelling and grammar are not drawn for the first letter part.

This is because layout object only points to the remainder part of
the text node excluding the first-letter part.

This CL handles the first-letter part in |DocumentMarkerController|,
|HighlightPainter| and |MarkerRangeMappingContext| to fix the issue.

Bug: 40301496
Change-Id: Idd73675f33bf31d9610d6b109f6632dc28fdc4d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5785328
Reviewed-by: Sanket Joshi <sajos@microsoft.com>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Sambamurthy Bandaru <sambamurthy.bandaru@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1349101}
NOKEYCHECK=True
GitOrigin-RevId: f30d9fddc4b561c3b14fea5692f7b1f1f0b516bd