DeferRenderTasksAfterInput: Fix feature param mapping
The "minimal-types" feature param is using the wrong constant, which is
configuration errors for the finch trial (and a DumpWithoutCrashing).
This wasn't caught by tests because we were looking up the param value
from the config based on the enum value. This CL fixes the bad mapping,
and it changes the unit tests to use the actual strings.

Fixed: 363133764
Change-Id: I91b068efc5177b2ce473003dadf091671988556f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5829622
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Scott Haseley <shaseley@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1349339}
NOKEYCHECK=True
GitOrigin-RevId: fffdc5fd96ecdc88e0fddbc3c880695fd9f1d1ad