[performance manager] Track WebRTC usage by pages and frames.
Tracking WebRTC usage is necessary to opt-out from freezing and
discarding tabs that are used for passively watching a videoconference
call (no video, audio or screen capture). For this use case, we don't
want to disconnect the tab when it's backgrounded because that event
can be visible by other participants.

Bug: 325954772
Change-Id: I9ce405658d8c4711feed476bb07493d5e68b3966
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5738793
Commit-Queue: Francois Pierre Doray <fdoray@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Patrick Monette <pmonette@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1349378}
NOKEYCHECK=True
GitOrigin-RevId: c536d7f5556f0b23019d7fd259bf6d2f6d4ae53b