Don't treat all inputs and textareas as dir=auto when their value changes.
Note that the added test in dir-auto-dynamic-changes.window.js doesn't
actually test the bug, which requires actual editing and not
manipulation of the value.  Thus I'm also adding a separate test with
testdriver.js to test a very simple case of the bug.

I'm putting this fix behind a flag because the code is quite old, just
in case there's some strange dependency on it.

Fixed: 362787713
Change-Id: Ib374d015332499a1532e045bdd413c26e110ba83
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5828471
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Commit-Queue: David Baron <dbaron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1349485}
NOKEYCHECK=True
GitOrigin-RevId: 4a344bab8c95e913a5156c964a78ece20b71ae24