[nesting] Cut off the outer rule body range at the first nested rule
With CSSNestedDeclarations, "bare" declarations that come between nested rules are wrapped in a CSSNestedDeclarations rule and are reported to CDP clients as standalone matches. The leading block of such declarations is still reported with the outer rule. We need to also adjust the `cssText` reported for the outer rule, so that it correctly  covers only those properties that are actually reported for it.

This fixes a bunch of issues in DevTools, which uses CSS.setStyleText to modify the cssText, ignoring the fact that CSS.setStyleText can't handle nested rules which would be part of the uncropped body.

Fixed: 41486635, 361477264, 328263458, 41487826
Change-Id: I62da536289ad70857ea9da105989765b3837ae66
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5822790
Commit-Queue: Philip Pfaffe <pfaffe@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1349877}
NOKEYCHECK=True
GitOrigin-RevId: b4b329c0e7ab0422dbee6d115f2041c6846f50cc