Make fieldset/legend UA stylesheet match spec more closely
The spec for fieldset/legend is here:

https://html.spec.whatwg.org/#the-fieldset-and-legend-elements

This reorganizes the UA stylesheet to match that, which only
practically makes one observable change: the border color changes
from `#C0C0C0` to `ThreeDFace`. Since there is no layout change
associated with this CL, it shouldn't be a risk for compat, because
there are no layout changes, and the color change is very slight.

ThreeDFace is defined at [1] and is #6B6B6B when the color scheme is
dark, or #EFEFEF when it is light. So this change makes the border
lighter than before in light mode, and darker than before in dark
mode.

[1] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/core/layout/layout_theme.cc;l=581;drc=5da8314cb77ca818d91111de8f33efa325a62315

Change-Id: Ib9f676e6ef9c9911a3146d4b038f5f144ffad0e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5806612
Auto-Submit: Mason Freed <masonf@chromium.org>
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1350216}
NOKEYCHECK=True
GitOrigin-RevId: d5adb3ebdcf66eb76115dee15fb3588b5cf03b9f