Factor EvaluationInput out of blink::Length
calc() expressions can be used in contexts other than where the result
is a Length, for example a channel value expression in relative color
syntax. It would make more sense for the input for resolving calc()
expressions not to be an inner class of Length, so this CL pulls it out.

Bug: 325309578
Change-Id: Ibc6bcb509f69e240786c8083b0cecc6c9b7a27ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5806114
Commit-Queue: Kevin Babbitt <kbabbitt@microsoft.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1350231}
NOKEYCHECK=True
GitOrigin-RevId: 9166feba61b495db3ba9e4ead6c279692f7a58ec