blink: Suppress scroll restoration for animated history navigations
If there is a visual transition for a history navigation, the browser
is displaying a preview of the final navigation state. If this
navigation is accompanied by a scroll offset change, this change
shouldn't be animated.

Fixed: 355494477
Change-Id: I2bd196204c36c3930542094320009291a3135a59
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5825293
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Commit-Queue: Vladimir Levin <vmpstr@chromium.org>
Auto-Submit: Khushal Sagar <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1350291}
NOKEYCHECK=True
GitOrigin-RevId: 38fc7fe405ab469b89641f1323a5f33a4e455376