run_web_tests.py: add option to send SIGTERM to test driver
A useful workflow is to run web tests with --wrapper="rr record", to
generate a debug recording of the test run. This works well if the test
completes normally or with an assertion failure; but if the test
crashes or times out, the test runner aggressively kills the test
process via SIGKILL, resulting in an incomplete recording.

This patch adds --kill-driver-with-sigterm, which causes SIGTERM to be
signalled to the test process prior to SIGKILL. This is sufficient to
get a complete recording from rr.

Also fixes a small bug in the treatment of --driver-kill-timeout-secs:
previously, if tracing was enabled on the test run, the value of
--driver-kill-timeout-secs was ignored in favor of a heuristic value
based on the test timeout (--timeout-ms). This patch causes the
greater of the two values to be used.

Change-Id: Ia337cf1877e32b2703f6db21b462658d0fadf761
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5818393
Reviewed-by: Weizhong Xia <weizhong@google.com>
Commit-Queue: Stefan Zager <szager@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1350327}
NOKEYCHECK=True
GitOrigin-RevId: 7b022b31ce9af5f090e3c3f4cfa51d50e4164d26