controlledframe: Add permissions policy
Controlled Frame is only available to IWAs and per our explainer
(https://github.com/WICG/controlled-frame) IWAs that intend to access
the API should request to use it via a permissions policy. This CL
adds that permissions policy, named "controlled-frame".

The new permissions policy is defined as a normal Blink permissions
policy and requires that the "ControlledFrame" runtime-enabled feature
is enabled. Finally, the permissions policy is only visible in an
isolated context, similar to other IWA APIs.

Since Controlled Frame is implemented as a wrapper around WebView, the
permissions policy check on both the renderer- and browser-sides is
in //extensions. The //extensions/renderer code to do this uses a
Blink public interface on WebLocalFrame. This CL modifies that class
to support checking IsFeatureEnabled() on the ExecutionContext object
it holds within Blink internal classes.

Bug: b/40191772, b/329720352
Change-Id: Ib83ac83fd466629623bec37c64c93c28656758e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5814831
Reviewed-by: Robbie McElrath <rmcelrath@chromium.org>
Reviewed-by: Ari Chivukula <arichiv@chromium.org>
Commit-Queue: Chase Phillips <cmp@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1350968}
NOKEYCHECK=True
GitOrigin-RevId: cbef0593d48ce7a6a5ffdfd8cd82f4250f5afea8