Set overflow: hidden for scroll snap test body
The test is flaky on Linux because the rightmost box in the line
occasionally exceeds the width of the viewport. This may be due to
calculations related to viewport width with scrollbars. Turning the
scrollbars off may help reduce flakiness. This is a speculative fix.

Bug: 363881422
Change-Id: I207a6f614ecb90c5cb11ae29dbd8e32943deaf8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5836267
Reviewed-by: Kevin Ellis <kevers@chromium.org>
Commit-Queue: Sahir Vellani <sahir.vellani@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1351033}
NOKEYCHECK=True
GitOrigin-RevId: 095c444231f9960d9f3d89ae1c1585c7d4464e0a