[rust png] Extracting `SkiaImageDecoderBase` from `GIFImageDecoder`.
This will help to reuse most of Skia <=> Blink integration in a
follow-up CL that integrates `SkPngRustCodec` into Blink.

This is a separate CL, because some additional tweaks of
`SkiaImageDecoderBase` are needed to make it work for PNG decoding.
Making those tweaks on top of this CL helps with reviewing/diffing the
tweaks from the follow-up CL.  (A WIP draft of the follow-up CL can be
found at https://crrev.com/c/5786777.)

Bug: chromium:356884491
Change-Id: I55e7de3d48adad39b3074aff962b9bbbedfa23b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5825156
Commit-Queue: Łukasz Anforowicz <lukasza@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1351168}
NOKEYCHECK=True
GitOrigin-RevId: b74f39b352a0fa4e36354239c22653b063356218