Roll emoji-segmenter, change mode to -F1
Switch emoji-segmenter to -F1 mode, see upstream commit [1].
This generates a potentially faster scanner code which avoids
binary search. This is the same mode that is used in HarfBuzz.

Documentation for -F1
   "-F1    (C/D/Ruby/C#/Go) Generate a faster flat table
    driven FSM by expanding action lists in the
    action execute code."

vs. a part of the documentation of the default, -T0:
    "...In  general,  the  table  driven FSM produces
     a smaller binary and requires a less expensive
     host language compile but results in slower code."

The generated code uses additional comparison operators
which are required to compile. Add the comparisons between
EmojiSegmentationCategory and uint8_t with the required casts.

Bug: chromium:40244183, chromium:41495057
Change-Id: I5179df0284eb7a0d174170003d8ec25790e2c28a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5837543
Reviewed-by: Munira Tursunova <moonira@google.com>
Commit-Queue: Dominik Röttsches <drott@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1351300}
NOKEYCHECK=True
GitOrigin-RevId: 38fe6db64d954b095c6974fd4872cb2a7c5c7a26