Re-enable fast/reflections/opacity-reflection-transform.html on Linux
This test got disabled in https://crrev.com/c/3202119.
It seems that the rebaseline-cl tool wasn't producing baselines for
the linux_layout_tests_layout_ng_disabled bot. This bot was removed in
https://crrev.com/c/4159905 and https://crrev.com/c/4152833. Thus, we
can now re-enable this test as it seems to be passing on all
platforms.

Comparing the commit history of the Linux [1] baseline with the
baseline for other other platforms (e.g. Windows [2]), we see that the
baseline for other platforms got updated multiple times since the test
got disabled on Linux. We therefore need to rebaseline the Linux test
as we re-enable it.

[1] https://crsrc.org/c/third_party/blink/web_tests/platform/linux/fast/reflections/opacity-reflection-transform-expected.png;bpv=1;bpt=0

[2]
https://crsrc.org/c/third_party/blink/web_tests/platform/win/fast/reflections/opacity-reflection-transform-expected.png;bpv=1;bpt=0

Bug: 40190804
Change-Id: Ia1ed84ecd3a73b725b3d71e61dbf4fb1548da72d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5815038
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Reviewed-by: Andres Ricardo Perez <andresrperez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1351435}
NOKEYCHECK=True
GitOrigin-RevId: b6e70dcf856c2ed0e6f21419301a7ddca060765e