Factor out CSS color function metadata for use in calc expressions
Currently, this metadata is used at parse time to determine which
channel names are valid for a given color function, and to resolve calc
expressions for color channels at parse time when the origin is known.
We'll also need this information to resolve calc expressions at
used-value time when the origin is `currentcolor`, so this CL lifts
the metadata to a location where it can be used outside of parsing.

Bug: 325309578
Change-Id: Ifc85e64d1d4b780f2dd1428f3429ac24cc5a904d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5714465
Reviewed-by: Fredrik Söderquist <fs@opera.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Kevin Babbitt <kbabbitt@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1351494}
NOKEYCHECK=True
GitOrigin-RevId: 51e5344f890c9da6e36a41abd7fd426ad3b75e8c