Propagate system preferred color-scheme changes to iframes
The resolved embedded preferred color-scheme was based on the modified
settings, so we need to store the old resolved value on the
HTMLFrameOwnerElement. Also, since the preferred color-scheme is not
stored on ComputedStyle, we can not detect the change in StyleDidChange
because StyleDidChange will not be called if the style does not
otherwise change.

Instead detect the change through DidRecalcStyle(). Changing the
preferred color-scheme settings causes a full document recalc which
means we will reach every element.

This issue was found investigating issue 364340032, but it does not fix
it.

Bug: 364340032
Change-Id: Ie3331f1cd1ad0d304e48fc5e07f11a133c6c91b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5836950
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1351890}
NOKEYCHECK=True
GitOrigin-RevId: bd9c4611bb930f46c5ee3cfa4a43372625da771f