IDB: Throw specific exception code for NotFound large value read errors
This CL adds a histogram to record instances of "Failed to read large
IndexedDB value" errors that also captures the FileErrorCode. Currently,
only two values of FileErrorCode are possible for IndexedDB -
kNotFoundErr and kNotReadableErr.

The DOMExceptionCode that is thrown to the application when this error
occurs is also updated to be kNotFoundError for the file not found case
and kDataError for all other cases (retaining current behavior). This
should help the application take the appropriate corrective action when
this error occurs.

Change-Id: I15ecf9d8dd75f638c68bf7c6720553046fd806c3
Bug: 362123231
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5831673
Reviewed-by: Ayu Ishii <ayui@chromium.org>
Reviewed-by: Evan Stade <estade@chromium.org>
Commit-Queue: Abhishek Shanthkumar <abhishek.shanthkumar@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1351910}
NOKEYCHECK=True
GitOrigin-RevId: d1ec7480f7af580e91623d129d373c1ef72f024b